Added an octree for faster view computation and fixed bugs.
The current code runs a test case which computes and displays one view of the tree. Change the first two arguments to tview.computeView() in glCanvas.cpp::display() if you want to render it from another angle.

+++ mesh.cpp
@@ -149,7 +149,7 @@ void Mesh::Load(const std::string &input_file) {
   int vert_count = 0;
   int vert_index = 1;
   Material *active_material = NULL;
-  std::vector<std::pair<int, int> > textures;
+  std::vector<std::pair<float, float> > textures;
 
   // read in each line of the file
   while (istr.getline(line,MAX_CHAR_PER_LINE)) {
